[DataView] Move outside pycore/
